Changes to computation of stabilization parameters.
This commit arose from the observation that the execution of MeshUpdate solver lead to failure of the TemperateIce solver, even when mesh update was zero.

The reason for this is twofold:

1) The stabilization parameters (% hk, specifically) are computed BEFORE MeshExtrude and StructuredMeshMapper, meaning they are computed on the 2D planar mesh.

2) DisplaceMesh triggers the recomputation of % hk, but actually, the value is the minimum edge length (the vertical, in this case), which isn't ideal for large aspect ratio elements with mostly advection in the long axis (most 3D glacier flow problems).

So, although the % hk computation was not updated, actually the 2D value was better for this problem. Thus, it's insufficient to simply update the stabilty parameters in StructuredMeshMapper. I've quite crudely changed the % hk parameter to equal the longest edge length.  This isn't necessarily ideal, but it fixes my problems for now.