Vector H in stabilisation for advection/diffusion
(advective heat equation, in my case)

This is code from Juha which hasn't made its way into the main repo yet. Since I don't fully understand it, and it's not my code, I don't intend to push this commit to the main repo.

In addition to Juha's work, I've also added the option to specify that the solver should be treated as steady state (Steady State Condition > 0).