Change to behaviour of ResultOutputSolve in terms of creating variable list
Instead of checking if the current mesh and ListMesh (mesh for which list was
last was created) are associated, we check if they have the same name.

This means that the list is preserved after a total remeshing, so long as the
mesh retains its original name.

The only potential downside here is that if a user had two different meshes
with the same name, this might cause a problem.