Added option UnfoundFatal to some ListGet functions
The following functions now accept the optional argument 'UnfoundFatal':

ListGetInteger, ListGetIntegerArray, ListGetLogical, ListGetString, ListGetReal, ListGetConstReal, ListGetConstRealArray, ListGetCReal, ListGetRealAtNode

If .TRUE. and the function fails to find the requested list item, a fatal error is called.

The idea of this, and the previous commit, is to make it easier and more concise to program defensively. Thanks to Rupert Gladstone for the idea.