Option to choose longest or shortest element length for stabilization
also added option in StructuredMeshMapper to recompute (or not) the stabilization parameters. This is a slightly complicated situation. Really, StructuredMeshMapper should *always* recompute the stabilization parameters, but the result of this (without also choosing the Longest Edge stabilization option) is that the (high aspect ratio) extruded mesh will have potentially insufficient stabilization, because by default the stabilization is computed using shortest element size.

Potential solution: have StructuredMeshMapper always recompute the stabilization parameter, and pass the argument that uses the longest element edge. If this is implemented, also update the tests.