Add SEMTEC Trafotek tests
Lots of development in the field of electromagnetism has been
done in SEMTEC project. This commit adds all the relevant tests
for the features that have been developed by Trafotek Oy.
The features include basic circuit equation fuctionality
with 3 kinds of coil formulations in 2D and 3D, new capabilities
in 2D magneto dynamics computations, and some basic homogenization
for stranded windings.

In total, there is 21 tests

* mgdyn2D_compute_average_b
* mgdyn2D_harmonic_anisotropic_permeability
* mgdyn2D_compute_bodycurrent
* mgdyn2D_scan_homogenization_elementary_solutions
* mgdyn2D_compute_complex_power
* circuits2D_harmonic_massive
* circuits2D_harmonic_stranded
* circuits2D_harmonic_stranded_explicit_coil_resistance
* circuits2D_harmonic_stranded_homogenization
* circuits2D_harmonic_foil
* circuits2D_transient_massive
* circuits2D_transient_stranded
* circuits2D_transient_foil
* circuits2D_scan_harmonics
* circuits_harmonic_massive
* circuits_harmonic_foil
* circuits_transient_stranded
* circuits_transient_massive
* circuits_harmonic_stranded
* circuits_harmonic_stranded_homogenization
* circuits_transient_foil