[NEW] Solver for elastic shell equations of Reissner-Naghdi type added
The shell solver creates a surface model by assuming that the surface normal
(director) is given at the background mesh nodes. The nodal director data
should be available via reading from file mesh.director located in the same
place as the standard mesh files or, alternatively, the user may provide a file
mesh.elements.data which should define the nodal director field associated with
the name 'director'.

This solver is STILL UNDER DEVELOPMENT and many changes are expected since some
possibilities of the solver are not yet fully utilized. Note the current
restrictions:

  -- Strain reduction operators (these are very essential) have been worked out
     for the lowest-order finite elements only
  -- Parallel file formats for the director data are still missing
  -- Postprocessing routines are also missing
  -- Some terms related to normal stretch field are still ignored in
     the expression for the strain energy (in order to make comparison with
     conventional shell models meaningful) but this is expected to change at
     a later point of time

The test cases that come with this commit provide verification against
reference results reported in the shell literature.