A benchmark problem for nonlinear shell analysis added as a test case
Solve nonlinear shell equations of Reissner-Naghdi type in the case of
a benchmark problem where a cantilever is subject to a shear force at an end.
The reference results can be found in Sze KY, Liu XH, Lo SH. Popular benchmark
problems for geometric nonlinear analysis of shells. Finite Elements in Analysis
and Design 2004, 40(11):1551-1569.