Apply a more standard space curve construction
The space curves needed in the surface reconstruction are now created in a more
standard manner. The Hermite interpolating function is now written directly
in terms of nodal tangent vectors that are created first from the geometry data.
Only the case of two background nodes per edge is supported currently.