An option to run the shell solver in parallel
The director data can now be generated by another solver module and the shell
solver can read this data provided the director field is given the name
'Director'. With this option the shell solver can also be run in parallel
without the burden of implementing parallel file formats for mesh.director or
mesh.elements.data.