Modify GetDomainEdge to allow getting the full top ring of nodes
This requires reordering the arguments so that EdgeMaskName is optional, hence changes in other solvers.

This is for a generalization of Calving3D (next commit) which doesn't require that a domain have 4 boundaries (inflow, left, calving, right) but any arbitrary number, so long as one of them is defined as the calving front.