Change the MeshGroup and other object naming to fix UNV names
Previously solid and boundary names could have '001' (or some other number)
added to the end of the UNV solid and boundary names and, hence, in Elmer body and
surface naming. Solved the issue by the renaming original solid objects,
relabeling the MeshGroup and using Labels in VTU export instead of Names.
This required also changes in the sif-files to take into account the corrected
naming.