Return the count of bubble DOFs solver-wise
If the sif file contains more than one solver section with each of them having
their own specification of the "Element" keyword, the function
GetElementNOFBDOFs may not return the number of bubbles that should be assigned
to the solver. With the optional argument Update = .TRUE. (this effect is now
the default), the correct solver-wise bubble count can be returned.