A quick fix for a computational geometry parameter
Reduce the size of a square that is tried to be embedded into the plain domain
associated with the shell element. A very irregular element shape may still lead
to an error in run time. It remains to figure out the precise size of a square
that can be embedded.