An option to apply the Piola transformation within FaceElementInfo
The function FaceElementInfo can now return the basis and the spatial divergence
of the basis as defined on the physical element. This simplifies writing a weak
formulation for an equation which is approximated with div-conforming
vector-valued finite elements (face elements). The face element examples within
the tests directory are also simplified by using this option.