Dirichlet BCs for div-conforming (face) finite elements in 2D
As a new feature Dirichlet BCs can be specified for div-conforming (face) finite
elements. The keyword syntax is chosen to be analogous to that used in
connection with curl-conforming (edge) finite elements. That is, if the Elmer
variable Q is approximated with face finite elements, the sif command

  Q {f} j = Real ...

specifies vector-valued data whose normal component is then used to integrate
the values of DOFs for vector-valued interpolation of the data. This commit
gives the implementation in 2D, so a 3-D version should still be made.

As a verification case a new test FaceElement2D_BCs is also added here. The
implementation of face element BCs has been checked to work for RT(0), ABF(0)
and BMD(1) in 2D.