Modify the labels of tests
- Add some missing "whitney" labels so that "ctest -L whitney" should run all
the tests for the models which utilize curl-conforming (edge) finite elements
to handle a version of the Maxwell equations. Unfortunately "whitney" is
semantically vague for that purpose as it indicates neither model nor type of
finite element accurately.
- Add a new label "vector_element" so that it can be used to run other tests
where vector-valued (edge or face) finite elements are employed.

To sum up, the sequence

 ctest -L whitney
 ctest -L vector_element

should now run all the tests related to div-conforming (face) or curl-
conforming (edge) finite elements.