Start writing a general-purpose mixed FEM solver for the Poisson equation
This commit gives an initial version of the solver which should evolve to be a
general-purpose mixed FEM solver for the Poisson equation -div(a grad u) = f.
An auxiliary unknown q is employed to approximate the vector q = a grad u by
using a div-conforming (face) finite element approximation. The assembly loop
over elements on boundary is still missing. This solver could have applications,
for example, in heat transfer, groundwater flow and electromagnetism.

An existing 3-D test case is also modified so that the new solver is now
employed instead of the special implementation in the test directory.