Natural BCs for the mixed formulation of the Poisson equation
The assembly loop over boundary elements is added so that the scalar variable
can be specified on a boundary in a weak manner (that is, as a natural/Neumann
BC in the context of the mixed formulation). The implementation is not yet
complete as this commit contains the implementation for the RT(0) approximation
only.

A simple test case is also added to verify that mixed boundary conditions (the
flux and the scalar field specified on complementary parts of the boundary) are
handled correctly.