Change the way how UMAT material models are included
This commit removes the file .../fem/src/UMAT.F90 and instead introduces a new
file ../fem/src/modules/UMATLib.F90 which contains both the template subroutine
for creating new material models in the umat form and some ready example
implementations. The old keywords "Use UMAT" and "User Defined UMAT" are now
made obsolite. To specify a material model in the umat form, one must now use
the keyword "UMAT Subroutine" in a material section in order to specify the file
(that can be compiled with an elmerf90 command before simulation) and to pick
the subroutine desired.

These changes remove the user's need to compile the whole solver of Elmer when
introducing a new material model (now a consistent version of elmerf90 is only
needed). For updated examples see also the directories .../fem/tests/UMAT_*