A correction for the A-V equations over a moving body
In a transient case where the mesh is transformed via a rigid motion the angular
velocity w of the rigid motion should be provided, so that a correction term
-(omega x A) can be generated in order to obtain a consistent time derivative
from the substantial time derivative which the time stepping machinery creates.
This term has been missing erroneously, but now the correction term is added to
the equations by reading the keywords "Angular velocity i" (i=1,2,3) from
Body Force sections.

The origin of the correction term is explained in the updated model documention
(Elmer Models Manual).

Some existing test cases are also revised so that the correction term is applied.