An additional option to set a positive normal direction
If NormalSolver is used to obtain a normal field for a surface model, a new
(bodywise given) keyword "Point on Negative Side" may be used to orient all
surface elements in the same way. This should help in getting all element
normals to point into the same side of the surface.