A subroutine for one-directional coupling of 3-D solids and 2-D shells
The 2-D shell solution can be constrained to obey a 3-D displacement field
which is expected to have the name "Displacement" and which should be solved
before the shell solver is executed. The implementation is not yet perfect as
it expects that 1) the nodes where the constraint is given must be listed by
using the "Target Nodes" keyword and 2) the shell director is not yet figured
out from the shell model data, so a redundant director specification should now
be given in a Boundary Condition section of the sif file.