Overhaul RenumberGDOFs to fix bugs relating to duplicate node nums
Previously assumed that, in any partition, any GDOFs which are no
longer used are globally available for use. However, if partition x
passed GDOF '100' to partition y, partition x believes 100 is
available, but that's not the case, leading to duplicate global node
numbering. This is now fixed (I believe).

Still need to rigorously test on other cases (e.g. when more nodes are
created than destroyed).