Overhaul of Calving3D_lset PlaneMesh methods - square mesh
Rather than using GetDomainEdge (which assumed certain mesh structure)
to find the top ring of nodes, we create a 2D grid mesh and pass that
to ProjectCalving. Then we use info about which PlaneMesh nodes didn't
hit the glacier to define the 'margins' of this PlaneMesh and pass
that information onto the IsoMesh.

This isn't quite working yet. Isomesh nodes/elements at the domain
edge are identified, but these are yet to be assigned front or
lateral.

Also, modifications to ValidateCrevassePaths etc should be made to
account for the new level set approach.