Unhardcode all (most?) parameters for levelset & aniso remeshing
Those relating to aniostropic remeshing live in the Material section
for now, because they ought to be solver agnostic. The calving
level-set cutting stuff is in the CalvingRemeshMMG solver section.

Also added USF_GlacierMeshMetric which computes an anisotropic metric
based on distance from the calving front. For now the z-component is constant.