Postprocessing for 2D magnetodynamics revised to include a Lorentz term
The computation of the electromagnetic field is changed to include
the term v x B. The computed E field should be inspected outside the source
region (the part where the current density is given) only, since the electric
field corresponding to the source is not taken into account. Therefore it is
better to consider the electric current density instead (by using "Calculate
Current Density = True"), which should be fully consistent.