Avoid misleading output or writing a misleading sif command
It is now possible to have "Timestepping Method = Bossak" when the problem is
of order 2 with respect to time derivatives. This commit doesn't have any
impact on the functionality (the Bossak method remains to be the only method for
the second-order problems), but this prevents the earlier "feature" of either
getting misleading output or giving a misleading sif command.