Revise the E-field and current density calculation when the velocity is given
The electromagnetic E-field calculation from the A-V solution is revised such
when the body velocity v is specified, the electric current density is obtained
in terms of E via multiplication with the electric conductivity. This change
conforms with the assumption of the charge q and force F being invariant
quantities (we define E = F/q). The E-field is thus computed as

E = -grad V  + v x curl A   (the steady state of evolutionary problem)

or

E = -DA/Dt - grad V  + w x A   (transient problem)

where w is the angular velocity and D/Dt is the substantial (total) time
derivative. For details see Elmer Models Manual.

An existing test case is modified so that the consistency between the computed
E-field and the current density field can be evidenced easily from the results,
together with a nicer visualization. Note that the revised version would
produce a different prediction of the E-field even if the material parameter of
the test case were left unaltered.