A new verification test for the electromagnetic rotation model
This is a simple verification case to check that a Faraday disc creates
the expected potential difference between the rim and and the axis.
As explained in the Readme file, this case is also instructive to understand
how Elmer treats Maxwell's equations in the case of a moving body using the
spatial description.