A verification case for the transient A-V model with a mesh rotation
This is a simple verification case to check that the transient model of a Faraday
disc creates the expected potential difference between the rim and and the axis.
This test verifies the magic interplay between Maxwell's equations rewritten
in terms the Lie derivatives, one-forms (the edge finite element approximation
of the vector potential) and the time stepping machinery of Elmer in combination
with a moving mesh. For additional information see the Readme file and
the updated documentation in the Elmer Models Manual.