Transient A-V models don't need explicit information about the velocity
Some time ago a wrong conjecture was made about the time derivative approximation
which arises in the case of a transient A-V model with a mesh rotation. A subtlety
is that in the case of one-forms (here the edge finite element approximation
of the vector potential) the time stepping machinery of Elmer in combination
with a moving mesh generates directly an approximation of the Lie derivative (not
the substantial derivative). For additional information see the updated
documentation in the Elmer Models Manual and also a new verification case
../fem/tests/mgdyn_faraday_disc_transient.