Fix a bug in SwitchMesh where multiple solvers share a variable (GL)
Not totally sure how this ever worked, to be honest. When two solvers
have the same variable (GroundedMask and GroundedMask Initialization),
the 2nd Solver's matrix won't be allocated in the 2nd
timestep (because it was shared with 1st solver) but it won't
necessarily be '.NOT. ASSOCIATED'. This means FreeMatrix attempts to
deallocate the Matrix_t (last line) even though it's already been
deallocced.

The fix: do the 2nd solver checking & repointing before the main
FreeMatrix. Seems to work.

I suspect this worked previously because often DEALLOCATE -> '.NOT. ASSOCIATED'
though this is never guaranteed.