The Helmholtz projection for curl-conforming fields
New utilities have been written as solvers to compute the Helmholtz projection
P(A) of a curl-conforming vector field A. The projection can be obtained as
P(A) = A - W where W is the curl-conforming field fitted to represent grad Phi,
with Phi being a H1-regular scalar field.

The solver HelmholtzProjector computes the scalar Phi, while the solver
RemoveKernelComponent computes W and then redefines the field A as A := A - W.

These utilities provide a way to remove the kernel component from a vector
potential A with respect to the operator curl. This can be performed for
example for each nonlinear iterate when the keyword "Nonlinear Post Solvers" is
used to active the utility solver calls. Another way to active the utility
solver calls in a post-solution manner could be to use the keyword
"Post Solvers" (provided the subroutine DefaultFinish is called within the
primary solver code).

Some examples are coming soon.