Examples of making the A-V potentials unique in an iterated manner
Alternate sif files for two existing tests are added (note that these alternatives
are not currently run under ctest). These depend on fresh utilities contained in
the module HelmholtzProjection.

The solution of the A-V system is not unique since gradient fields belong to the
kernel of curl. The ambiguity could be resolved by augmenting the A-V system so
as to impose the divergence-free constraint on A in a weak manner. This would
come with an expense of increasing the number of unknown fields.

The new example sif files use an alternate approach where the vector potential
variable is made unique in an iterated manner (the nonlinear iteration is
employed for this purpose) by performing the Helmholtz projection after solving
the A-V solution candidate. The potential benefits of this approach have not yet
been investigated thoroughly, so this is still somewhat experimental work.