Circuits: Implement voltage factor
Here we add a so called circuit voltage factor for circuit
components. The typical usage could be taking into accoung
symmetries. For example if only a half of a winding is solved
then one can add a voltage factor 0.5 to a components as
follows:

Component 1 :: Circuit Equation Voltage Factor = Real 0.5

The effect of this is that the voltage is multiplied with 0.5
in the circuit equation, that is, only one half of the real voltage
is in the circuit equation. This means that the resulting voltage
is twice the voltage without the factor.

All the tests have been modified to use this factor.