Another example of making the A-V potentials unique in an iterated manner
An alternate sif file for the existing test is added (note that this alternative
is not currently run under ctest). The alternative makes the vector potential
variable unique in an iterated manner (in connection with the nonlinear
iteration) by performing the Helmholtz projection after solving the A-V
solution candidate. The nonlinear iteration thus combines two tasks. It resolves
the nonlinearity due to the B-H law and makes the vector potential unique.