A generic solver to add node-wise springs to structural models
This new solver can be called as an additional assembly routine by using the
keyword "Assembly Solvers" in the solver section associated with the model
which is modified to have the springs. Element-wise spring constraints over
higher-dimensional entities must still be handled in the primary solver code.
This solver assumes that the places of the springs are listed using the keyword
"Target Nodes". It should be possible to utilize this solver in connection with
several structural models:

- linear elasticity
- nonlinear elasticity
- shells (currently only with the linearized version, with  "Large Deflection=False")
- plates (untested)
- beams (untested)