A new test: apply mortar BCs and make the A-V potentials unique simultaneously
This is a special model of the Faraday wheel to test the mortar FE approximation
when the A-V solution is enforced to be unique at the same time. The vector
potential variable is made unique in an iterated manner (the nonlinear iteration
is employed for this purpose) by performing the Helmholtz projection after
solving the A-V solution candidate.