Possibility to add node-wise masses to structural models
The utility SpringAssembly can now be called as an additional assembly routine
to add node-wise (generalized) masses to structural models using the keyword
"Mass k". Here it is assumed that the places of the masses are listed using the
keyword "Target Nodes". It should be possible to utilize this solver in connection
with several structural models:

    - linear elasticity
    - nonlinear elasticity
    - shells
    - plates
    - beams