Fix values of Solver % Def_Dofs to get high-order quadratures
When a MATC function had been used to define the order of p-basis, the function
isActivePElement didn't return a correct value. Solver % Def_Dofs is updated in
an attempt to resolve the trouble.

The supposed arguments in the element definition of the form

Element = "p:%a_matc_function"

are also changed, so that the first component of the 4-tuple given as the argument
for the MATC function is the identifier of the body while the remaining components
are still the coordinates of the element mid-point. This should offer a way to define
values of p in a body wise manner. This can be reverted if the original version
is thought to be better.