Combined analysis of 2-D shells and 1-D beams
The first version of an enhancement which allows a shell model to have one-
dimensional beam sections. At the moment the combined model can be assembled in
the case of linear analysis only (Large Deflection = False).

All parameters of the beam model are read from material sections. In particular
one of the principal directions of the beam cross section (the y3-axis associated
with the basis vector e3) can be given by specifying the value of the keyword
"Director". The beam stiffness matrix is manipulated in such a way that
the resultant couple (moment) vector cannot have a component in span{e3}. Since
the shell model does not have a rotational DOF around its director, it is natural
to construct the model so that the shell and beam directors agree when the domains
of the two models intersect.