Preliminary support for drilling DOFs within ShellSolver
If the command "Drilling DOFs = True" is given for the shell solver, the sixth
field variable is taken to be a drilling rotation instead of the normal strain
in the direction of the shell director. Drilling DOFs do not support all options
and alters the meaning of all rotational DOFs/BCs, since the rotation variables
Beta used locally are related to the corresponding global variables Theta via
the rotation Beta = d x Theta. Therefore a sif generated for the normal strain
formulation may not work with drilling DOFs without reconsidering BCs. Using
drilling DOFs also break current coupling procedures for more advanced simulations.

It seems that drilling DOFs may have a relative merit when the middle-surface is
not smooth and the director data has large jumps.