Try to harmonize the use of keywords for the AV solvers
The AV solvers for 2D and 3D cases have supposed different ways to specify the
magnetic anisotropy. Now the AV solvers in 3D should also allow the user to specify
the reluctivity by using a 3-tuple when the coordinate lines are the principal axes
of the reluctivity tensor, i.e. just the diagonal entries of the reluctivity
tensor may be given instead of all nine entries.