Implement W vector for circuit components
This is a new feature for giving the wire density as
a vector. Previously this has been given as a potential.
Potential is an easy solution as it is gotten easily
from Poisson's equation with good properties.
The draw back is that the potential cannot be used
in full coils.

Here the implementation assumes a wire density vector
variable that is directly used as the wire density vector
if "Use Coil W Vector = True" (see circuits_transient_stranded_wvector).
Currently the name of the variable must be "W Vector E".
A keyword "W Vector Variable Name" will be added later.
Moreover, the W Vector feature works only for transient stranded coil
at the moment.