Commentaries on def_dofs and a small correction
All entries of the def_dofs array of the solver structure are now copied when
the solver section defines the mesh. The def_dofs arrays are now given as
arguments when calling the subroutine GetDefs. Hopefully it is thus clearer
which sections of the function LoadModel write or use the def_dofs arrays.