Fixes for MATC-dependent element definitions for p-FEM
- Investigate a right string when checking for MATC-dependent element definitions
- Have the right element string when processing several elements within the same body
- A partial fix to avoid creating an erroneous p-distribution. The updated code
corrects this when the MATC-dependent element definition is given in Equation
section so that all active solvers use the same p-distribution.

In addition, call the subroutine GetMaxDefs which has limited abilities in processing
element commands only when it is strictly necessary (the subroutine GetDefs has
better support for element commands).