MPI_Init only if required
When using Elmer as a library from a software that already runs MPI
calling MPI_Init again is not working, so to enable this check if MPI is
already initialized and skip Init if it is already there. In this case
we also don't want to call MPI_Finalize as the other software should
take care of that.