A new keyword to guide the shell reparametrization
If it is known that a body represents a part of a sphere, the reparametrization
of the shell mid-surface can be guided by giving the command "Spherical Body =
True" in the body section. Then the shell solver doesn't try to decide umbilical
regions by a computational criterion. It seems that this explicit guidance may
give much better results if the mesh is not regular (for example high aspect
ratios or irregular element shapes).