A 4-node plate element based on the full linked interpolation
One can now choose a 4-node formulation which starts by adding 5 hierarchic bubble
functions to the deflection and then finding their coefficients by a linked
interpolation strategy. The shear correction trick is also applied. A nice feature
of this formulation is that it connects the formulation to approximating the rotation
in the (curl-conforming) ABF_0 space. This formulation may have a relative merit
when the mesh is not regular, but otherwise the solution may be expected to be close
to that obtained with the stabilized MITC4.