[NEW/FIX] Full geometric nonlinearity with low-order shell elements
This commit alters the shell model which is used in nonlinear analysis with low-
order finite elements. Previously the zeroth-order terms in the expansions of
strain components in powers of the thickness coordinate were only treated as
nonlinear, but now the first-order terms are also taken to be nonlinear. This
prevents some anomalies which appear to be possible with the previously
implemented shell model. In the case of strong nonlinearity a notable change in
the solution may thus be possible. In order to switch to the earlier shell model,
the command "Nonlinear Bending Strains = False" can be used.

A smaller change is that the linearized bending strains are also made fully
consistent by adding terms which depend on the expressions of the membrane
strains. This should not change the shell model in an essential way.