Hidden  mesh, compass, etc. during double clicking
In mouseDoubleClickEvent(QMouseEvent), surface mesh, volume mesh, compass and sharp edges are hidden to avoid segmentation fault. This is just a kind of temporary measures. I don't understand why this segmentation fault occurs, but it didn't happen when meshes, compass and sharp edges were hidden as far as I know.