[NEW] Higher-order shell elements backgrounded by a standard surface mesh
General shell geometries can now be handled by higher-order shell finite
elements. The user must supply a standard low-order surface mesh embedded in 3-D
space together with the director data, so that the shell solver may internally
attempt an improved surface reconstruction. At the moment the surface
reconstruction is based on mapped polynomials of degree 3, so the accuracy is
not expected to be optimal for all element orders, since geometric errors
eventually dominate the error. Numerical over-stiffness due to locking can
however be treated effectively by using finite elements of a sufficiently high
order, so even a non-optimal approximation combination might improve the solution
if the dominating error is caused by locking. Note that the high-order version
does not apply any strain reduction operators to alleviate locking, so with this
approach increasing the element order is the only way to handle locking.

To enable this approach the following keyword commands are needed at the moment:

  Cartesian Formulation = Logical True
  Skip Surface Reconstruction = Logical False
  Element = ...

This has preliminarily been tested to work, but to create a quad of degree 3,
consider giving "Element = p:3 b:4" instead of the serendipity element definition
"Element = p:3" (cf. the Readme file of the new test
Shell_OpenHemisphere_High_Order_Blending).