Economic 12-point quadrature for p-quads is faulty (?)
The avoidance of integrating with this rule makes the shell solver to produce
reasonable results with the element definition "p:3". Peculiarly this rule
nevertheless seems to work for lower-order integrands.

The function GaussPoints is now modified so that we never call the 12-point rule.
This appears to fix the shell solver issue left open in the commit 212d335.