[NEW] High-order visualization for the p-version of FEM
ResultOutputSolve can now generate files in the vtu format which, in contrast to the
earlier possibility of saving a p-solution only at the nodes of background mesh,
enable high-order visualization of p-FEM results. The polynomial degree k related to
the visualization can be specified by using the command

   Lagrange Element Degree = k

which is given in the section controlling ResultOutputSolve. In priciple it is
reasonable to choose k to be the same as the polynomial order of the p-solution (defined
by a command "Element = p:..."), but the two values can be chosen independently.
In 2D (quads and triangles) and in 3D for hexahedra this option works for the values of
k up to 8. Otherwise the functionality is not finalized yet and is limited to lower values
of k. At the moment tetrahedra and prisms allow values up to 4, while pyramids allow at
most the second-order visualization. Some node indexing routines have to be revised to
allow more complete support in 3D. This is expected to be addressed in future.

For examples see cases which are under tests directory and use the keyword "Lagrange
Element Degree".

The reader software of vtu files has to use a relatively new version of VTK (for example
ParaView versions 5.5 or newer), so that the cell types which support high-order
Lagrange interpolation are included.

This commit merges and resolves the branch 'p_to_Lagrange' of
https://github.com/ElmerCSC/elmerfem into devel.
Co-authored-by: Peter Raback <peter.raback@csc.fi>