Implement London equations to MagnetoDynamics2D
The circuit driven vector potential formulation with the London equations can
be solved by solving three equations:
  1. (v curl a, curl a') + (Lambda^{-1} a, a') + Phi (Lambda^{-1} grad phi0, a') = 0
  2. (Lambda^{-1} a, a') + Phi (Lambda^{-1} grad phi0, grad phi') = I
  3. d(Phi)/dt - v = 0

where Lambda = mu_0 lambda_l^2 and lambda_l is the london penetration
depth.

This commit implements the eqns. 1 and 2. The third equation has to be
written by the user. Two tests are added as an example and consistency.

Test cases:
  1. fem/tests/circuits2D_harmonic_london
  2. fem/tests/circuits2D_transient_london

Notice that the london equation for a material without component
equation is activated by giving a material parameter "London Lambda" and
for the component it is activated if "London Equations = Logical True"
is set in the component section.

The tests are the same case in time harmonic and transient domains.
Note that the transient case depends on the initial condition of the
supercurrents and thus the magnetic flux is shifted where as the
time harmonic case is symmetric.