New construct "variable_name {n} ..." for giving BCs
If an element has nodal DOFs created with an element definition

   Element = "n:k ..."

where k > 0, BCs may be given by using a part {n} in the BC definition.
For example, if we have

   Variable = E
   Variable DOFs = 1
   Element = "n:1 ..."

then a BC definition

   E {n} = Real ...

should also work. On the other hand, if

   Variable = E
   Variable DOFs = 1
   Element = "n:2 ..."

then we may give

   E {n} 1 = Real ...
   E {n} 2 = Real ...

Finally, if we the variable is defined to have components and the component
names are created for example from

   Variable = E[E Re:1 E Im:1]
   Element = "n:2 ..."

we may define

  E re {n} 1 = Real ...
  E re {n} 2 = Real ...
  E im {n} 1 = Real ...
  E im {n} 2 = Real ...

NOTE: This is still simplistic implementation and lacks many options which work
      in the case of standard BCs for nodal (Lagrange) interpolation
      approximations.