A gauged A-V formulation for the time-harmonic EM waves
Defining two nodal fields with an element definition "Element = n:2 ..."
activates this formulation. NOTE: At the moment there seems to be a problem in
creating the vtu output of nodal fields (with Calculate Nodal Fields = True),
but the elementwise approximations seem to be as expected in a simple test case
for which a sif file is also added.