Make IBC component equation compatible with a-part
Lets use impedance at the component equation. This should be compatible
with the impedance boundary condition. When we don't use circuits, we
seem to assume a current sheet with thickness of skin depth. But when we
do that at the compoent equations, we don't get a coherent current (when
comparing conventional massive conductor with circuits). Let's use
impedance then and see if that works better. I am wondering the
correctness of the current sheet thinking in the case of IBC without
circuits.