Fix SUBROUTINE FindMeshEdges3D
High-order visualization for p-elements in 3D (with the keyword Lagrange Element Degree) has been broken after the commit 36093a63cd1288b2a1592f07486fb2b996598b3c. The data object ElementIndex associated with the mesh edges lost in the commit is again created so that the high-order visualization should now work.

In addition, another (unrelated) correction is made so that the edges corresponding to the faces of a tetrahedron are again indexed.