Add CI rule that triggers once a week with the full test suite.
Currently, this is basically a copy of the build rules that trigger on
pull requests or pushes. The only difference is that the trigger is on a
weekly schedule and the full set of tests are run.

Mid-term, more tests could be run for different configurations on PRs or
pushes (e.g., build tests with bundled or external dependencies, with
some enabled or disabled features, with different compilers, on different
platforms, ...) while the full test suite is only run with one or a
different subset of configurations.

That will be easier if we start with two sets of (currently basically
identical) rules and let them spread in different directions from here
without having to care about the other rules.